Added parsing of conditionals
