+++ able-script-test/parse_test.able
@@ -1,1 +1,3 @@
-bff a {+++<<>>>[]]]][[]]}
+if (true) {
+    var a = 3;
+}
