+++ src/parser/item.rs
@@ -1,6 +1,9 @@
 use crate::variables::Value;
 
 #[derive(Debug, Clone)]
+pub struct Iden(pub String);
+
+#[derive(Debug, Clone)]
 pub enum Expr {
     VariableDeclaration {
         iden: String,
@@ -14,5 +17,11 @@ pub enum Expr {
         iden: String,
         body: String,
     },
+    If {
+        cond: Box<Expr>,
+        body: Vec<Expr>,
+    },
+
     Literal(Value),
+    Melo(Iden),
 }
